Separate macro expansion tests that expect errors
