+++ graphql_compiler/tests/test_helpers.py
@@ -7,6 +7,7 @@ from graphql import parse
 from graphql.utils.build_ast_schema import build_ast_schema
 import six
 
+from ..macros import create_macro_registry, register_macro_edge
 from ..debugging_utils import pretty_print_gremlin, pretty_print_match
 from ..query_formatting.graphql_formatting import pretty_print_graphql
 from ..schema import insert_meta_fields_into_existing_schema
@@ -245,3 +246,113 @@ def construct_location_types(location_types_as_strings):
         location: schema.get_type(type_name)
         for location, type_name in six.iteritems(location_types_as_strings)
     }
+
+
+def get_test_macro_registry():
+    schema = get_schema()
+    macro_registry = create_macro_registry()
+    type_equivalence_hints = {
+        schema.get_type('Event'): schema.get_type('EventOrBirthEvent'),
+    }
+
+    valid_macros = [
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_GrandparentOf") {
+                out_Animal_ParentOf {
+                    out_Animal_ParentOf @macro_edge_target {
+                        uuid
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_RichSiblings") {
+                in_Animal_ParentOf {
+                    net_worth @tag(tag_name: "parent_net_worth")
+                    out_Animal_ParentOf @macro_edge_target {
+                        net_worth @filter(op_name: ">", value: ["parent_net_worth"])
+                        out_Animal_BornAt {
+                            event_date @filter(op_name: "<", value: ["%birthday"])
+                        }
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Location @macro_edge_definition(name: "out_Location_Orphans") {
+                in_Animal_LivesIn @macro_edge_target {
+                    in_Animal_ParentOf @filter(op_name: "has_edge_degree", value: ["$num_parents"])
+                                       @optional {
+                        uuid
+                    }
+                }
+            }
+        }''', {
+            'num_parents': 0,
+        }),
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_RichYoungerSiblings") {
+                net_worth @tag(tag_name: "net_worth")
+                out_Animal_BornAt {
+                    event_date @tag(tag_name: "birthday")
+                }
+                in_Animal_ParentOf {
+                    out_Animal_ParentOf @macro_edge_target {
+                        net_worth @filter(op_name: ">", value: ["net_worth"])
+                        out_Animal_BornAt {
+                            event_date @filter(op_name: "<", value: ["%birthday"])
+                        }
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_AvailableFood") {
+                out_Animal_LivesIn {
+                    in_Entity_Related {
+                        ... on Food @macro_edge_target {
+                            uuid
+                        }
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Animal @macro_edge_definition(name: "invalid_out_Animal_AvailableFood") {
+                out_Animal_LivesIn {
+                    in_Entity_Related @macro_edge_target {
+                        ... on Food {
+                            uuid
+                        }
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_NearbyEvents") {
+                out_Animal_LivesIn {
+                    in_Entity_Related @macro_edge_target {
+                        ... on Event {
+                            uuid
+                        }
+                    }
+                }
+            }
+        }''', {}),
+        ('''{
+            Animal @macro_edge_definition(name: "out_Animal_NearbyEntities") {
+                out_Animal_LivesIn {
+                    in_Entity_Related {
+                        ... on Entity @macro_edge_target {
+                            uuid
+                        }
+                    }
+                }
+            }
+        }''', {}),
+    ]
+
+    for graphql, args in valid_macros:
+        register_macro_edge(macro_registry, schema, graphql, args, type_equivalence_hints)
+    return macro_registry
+
