+++ graphql_compiler/tests/test_macro_expansion.py
@@ -3,118 +3,8 @@ import unittest
 
 import pytest
 
-from ..exceptions import GraphQLCompilationError
-from ..macros import create_macro_registry, perform_macro_expansion, register_macro_edge
-from .test_helpers import compare_graphql, get_schema
-
-
-def get_test_macro_registry():
-    schema = get_schema()
-    macro_registry = create_macro_registry()
-    type_equivalence_hints = {
-        schema.get_type('Event'): schema.get_type('EventOrBirthEvent'),
-    }
-
-    valid_macros = [
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_GrandparentOf") {
-                out_Animal_ParentOf {
-                    out_Animal_ParentOf @macro_edge_target {
-                        uuid
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_RichSiblings") {
-                in_Animal_ParentOf {
-                    net_worth @tag(tag_name: "parent_net_worth")
-                    out_Animal_ParentOf @macro_edge_target {
-                        net_worth @filter(op_name: ">", value: ["parent_net_worth"])
-                        out_Animal_BornAt {
-                            event_date @filter(op_name: "<", value: ["%birthday"])
-                        }
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Location @macro_edge_definition(name: "out_Location_Orphans") {
-                in_Animal_LivesIn @macro_edge_target {
-                    in_Animal_ParentOf @filter(op_name: "has_edge_degree", value: ["$num_parents"])
-                                       @optional {
-                        uuid
-                    }
-                }
-            }
-        }''', {
-            'num_parents': 0,
-        }),
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_RichYoungerSiblings") {
-                net_worth @tag(tag_name: "net_worth")
-                out_Animal_BornAt {
-                    event_date @tag(tag_name: "birthday")
-                }
-                in_Animal_ParentOf {
-                    out_Animal_ParentOf @macro_edge_target {
-                        net_worth @filter(op_name: ">", value: ["net_worth"])
-                        out_Animal_BornAt {
-                            event_date @filter(op_name: "<", value: ["%birthday"])
-                        }
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_AvailableFood") {
-                out_Animal_LivesIn {
-                    in_Entity_Related {
-                        ... on Food @macro_edge_target {
-                            uuid
-                        }
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Animal @macro_edge_definition(name: "invalid_out_Animal_AvailableFood") {
-                out_Animal_LivesIn {
-                    in_Entity_Related @macro_edge_target {
-                        ... on Food {
-                            uuid
-                        }
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_NearbyEvents") {
-                out_Animal_LivesIn {
-                    in_Entity_Related @macro_edge_target {
-                        ... on Event {
-                            uuid
-                        }
-                    }
-                }
-            }
-        }''', {}),
-        ('''{
-            Animal @macro_edge_definition(name: "out_Animal_NearbyEntities") {
-                out_Animal_LivesIn {
-                    in_Entity_Related {
-                        ... on Entity @macro_edge_target {
-                            uuid
-                        }
-                    }
-                }
-            }
-        }''', {}),
-    ]
-
-    for graphql, args in valid_macros:
-        register_macro_edge(macro_registry, schema, graphql, args, type_equivalence_hints)
-    return macro_registry
+from ..macros import perform_macro_expansion
+from .test_helpers import compare_graphql, get_schema, get_test_macro_registry
 
 
 class MacroExpansionTests(unittest.TestCase):
@@ -198,57 +88,6 @@ class MacroExpansionTests(unittest.TestCase):
             self.schema, self.macro_registry, query, args)
         compare_graphql(self, expected_query, expanded_query)
         self.assertEqual(expected_args, new_args)
-
-    # TODO(bojanserafimov): Move test_macro_expansion_errors.py to test_macro_validation.py
-    # Move this function and all the others to test_macro_expansion_errors.py
-    @pytest.mark.skip(reason='not implemented')
-    def test_macro_edge_duplicate_edge_traversal(self):
-        query = '''{
-            Animal {
-                out_Animal_BornAt {
-                    uuid
-                }
-                out_Animal_RichYoungerSiblings {
-                    uuid
-                }
-            }
-        }'''
-        args = {}
-
-        with self.assertRaises(GraphQLCompilationError):
-            perform_macro_expansion(self.schema, self.macro_registry, query, args)
-
-    @pytest.mark.skip(reason='not implemented')
-    def test_macro_edge_target_coercion_invalid_1(self):
-        query = '''{
-            Animal {
-                out_Animal_AvailableFood {
-                   ... on Species {
-                       @output(out_name: "species")
-                   }
-                }
-            }
-        }'''
-        args = {}
-
-        with self.assertRaises(GraphQLCompilationError):
-            perform_macro_expansion(self.schema, self.macro_registry, query, args)
-
-    @pytest.mark.skip(reason='not implemented')
-    def test_macro_edge_invalid_coercion_2(self):
-        query = '''{
-            Animal {
-                out_Animal_NearbyEvents {
-                   ... on Entity {
-                       @output(out_name: "event")
-                   }
-                }
-            }
-        }'''
-        args = {}
-
-        with self.assertRaises(GraphQLCompilationError):
-            perform_macro_expansion(self.schema, self.macro_registry, query, args)
 
     @pytest.mark.skip(reason='not implemented')
     def test_macro_edge_target_coercion_1(self):
@@ -388,17 +227,3 @@ class MacroExpansionTests(unittest.TestCase):
             self.schema, self.macro_registry, query, args)
         compare_graphql(self, expected_query, expanded_query)
         self.assertEqual(expected_args, new_args)
-
-    @pytest.mark.skip(reason='not implemented')
-    def test_macro_edge_nonexistent(self):
-        query = '''{
-            Animal {
-                out_Animal_GrandparentOf {
-                    @output(out_name: "grandkid")
-                }
-            }
-        }'''
-        args = {}
-
-        with self.assertRaises(GraphQLCompilationError):
-            perform_macro_expansion(self.schema, self.macro_registry, query, args)
